melhoria no codigo
