+++ src/services/makeAPICallPost.js
@@ -1,0 +1,12 @@
+export default async function makeAPICallPost(url, body) {
+  console.log("ok");
+  const response = await fetch(url, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(body),
+  });
+  const data = await response.json();
+  return data;
+}
